upload
Made to funtion with WP all import and Woocommerce. 
In october 2021 there was a change of the currency in venezuela. It was a law that all the stores had to show the prices in the new currency and the old one AT THE SAME TIME. We had no time to change the currency plugin so I created a custom field in the woocommerce products to storage the new price calculated by this function and then just displayed it later with a woocommerce's hook